+++ config/spotbugs/exclude.xml
@@ -262,7 +262,7 @@
 
     <!-- mongocrypt -->
     <Match>
-        <Class name="com.mongodb.crypt.capi.CAPI$cstring"/>
+        <Class name="com.mongodb.internal.crypt.capi.CAPI$cstring"/>
         <Bug pattern="NM_CLASS_NAMING_CONVENTION"/>
     </Match>
