+++ driver-core/src/main/com/mongodb/internal/capi/MongoCryptHelper.java
@@ -24,10 +24,10 @@ import com.mongodb.MongoClientException;
 import com.mongodb.MongoClientSettings;
 import com.mongodb.MongoConfigurationException;
 import com.mongodb.client.model.vault.RewrapManyDataKeyOptions;
-import com.mongodb.crypt.capi.MongoCryptOptions;
 import com.mongodb.internal.authentication.AwsCredentialHelper;
 import com.mongodb.internal.authentication.AzureCredentialHelper;
 import com.mongodb.internal.authentication.GcpCredentialHelper;
+import com.mongodb.internal.crypt.capi.MongoCryptOptions;
 import com.mongodb.lang.Nullable;
 import org.bson.BsonDocument;
 import org.bson.BsonDocumentWrapper;
