+++ driver-reactive-streams/src/main/com/mongodb/reactivestreams/client/internal/crypt/KeyManagementService.java
@@ -24,7 +24,6 @@ import com.mongodb.ServerAddress;
 import com.mongodb.connection.AsyncCompletionHandler;
 import com.mongodb.connection.SocketSettings;
 import com.mongodb.connection.SslSettings;
-import com.mongodb.crypt.capi.MongoKeyDecryptor;
 import com.mongodb.internal.TimeoutContext;
 import com.mongodb.internal.TimeoutSettings;
 import com.mongodb.internal.connection.AsynchronousChannelStream;
@@ -33,6 +32,7 @@ import com.mongodb.internal.connection.OperationContext;
 import com.mongodb.internal.connection.Stream;
 import com.mongodb.internal.connection.StreamFactory;
 import com.mongodb.internal.connection.TlsChannelStreamFactoryFactory;
+import com.mongodb.internal.crypt.capi.MongoKeyDecryptor;
 import com.mongodb.internal.diagnostics.logging.Logger;
 import com.mongodb.internal.diagnostics.logging.Loggers;
 import com.mongodb.internal.time.Timeout;
