+++ mongodb-crypt/src/main/com/mongodb/internal/crypt/capi/JULLogger.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.mongodb.crypt.capi;
+package com.mongodb.internal.crypt.capi;
 
 import java.util.logging.Level;
