+++ mongodb-crypt/src/main/com/mongodb/internal/crypt/capi/Logger.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.mongodb.crypt.capi;
+package com.mongodb.internal.crypt.capi;
 
 /**
  * Not part of the public API
