+++ mongodb-crypt/src/main/com/mongodb/internal/crypt/capi/MessageDigestCallback.java
@@ -15,20 +15,20 @@
  *
  */
 
-package com.mongodb.crypt.capi;
+package com.mongodb.internal.crypt.capi;
 
-import com.mongodb.crypt.capi.CAPI.cstring;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_binary_t;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_hash_fn;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_status_t;
+import com.mongodb.internal.crypt.capi.CAPI.cstring;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_binary_t;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_hash_fn;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_t;
 import com.sun.jna.Pointer;
 
 import java.security.MessageDigest;
 
-import static com.mongodb.crypt.capi.CAPI.MONGOCRYPT_STATUS_ERROR_CLIENT;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_set;
-import static com.mongodb.crypt.capi.CAPIHelper.toByteArray;
-import static com.mongodb.crypt.capi.CAPIHelper.writeByteArrayToBinary;
+import static com.mongodb.internal.crypt.capi.CAPI.MONGOCRYPT_STATUS_ERROR_CLIENT;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_set;
+import static com.mongodb.internal.crypt.capi.CAPIHelper.toByteArray;
+import static com.mongodb.internal.crypt.capi.CAPIHelper.writeByteArrayToBinary;
 
 class MessageDigestCallback implements mongocrypt_hash_fn {
