+++ mongodb-crypt/src/main/com/mongodb/internal/crypt/capi/MongoCrypt.java
@@ -16,7 +16,7 @@
  */
 
 
-package com.mongodb.crypt.capi;
+package com.mongodb.internal.crypt.capi;
 
 import org.bson.BsonDocument;
