+++ mongodb-crypt/src/main/com/mongodb/internal/crypt/capi/MongoCryptContextImpl.java
@@ -15,33 +15,34 @@
  *
  */
 
-package com.mongodb.crypt.capi;
+package com.mongodb.internal.crypt.capi;
 
-import com.mongodb.crypt.capi.CAPI.mongocrypt_binary_t;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_t;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_kms_ctx_t;
+import com.mongodb.crypt.capi.MongoCryptException;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_binary_t;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_t;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_kms_ctx_t;
 import org.bson.BsonDocument;
 import org.bson.RawBsonDocument;
 
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_binary_destroy;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_binary_new;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_destroy;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_finalize;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_kms_done;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_mongo_done;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_mongo_feed;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_mongo_op;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_next_kms_ctx;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_provide_kms_providers;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_state;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_status;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_code;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_destroy;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_message;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_new;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_t;
-import static com.mongodb.crypt.capi.CAPIHelper.toBinary;
-import static com.mongodb.crypt.capi.CAPIHelper.toDocument;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_binary_destroy;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_binary_new;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_destroy;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_finalize;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_kms_done;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_mongo_done;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_mongo_feed;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_mongo_op;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_next_kms_ctx;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_provide_kms_providers;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_state;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_status;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_code;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_destroy;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_message;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_new;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_t;
+import static com.mongodb.internal.crypt.capi.CAPIHelper.toBinary;
+import static com.mongodb.internal.crypt.capi.CAPIHelper.toDocument;
 import static org.bson.assertions.Assertions.isTrue;
 import static org.bson.assertions.Assertions.notNull;
