+++ mongodb-crypt/src/main/com/mongodb/internal/crypt/capi/MongoCryptImpl.java
@@ -15,13 +15,14 @@
  *
  */
 
-package com.mongodb.crypt.capi;
+package com.mongodb.internal.crypt.capi;
 
-import com.mongodb.crypt.capi.CAPI.cstring;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_t;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_log_fn_t;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_status_t;
-import com.mongodb.crypt.capi.CAPI.mongocrypt_t;
+import com.mongodb.crypt.capi.MongoCryptException;
+import com.mongodb.internal.crypt.capi.CAPI.cstring;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_t;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_log_fn_t;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_t;
+import com.mongodb.internal.crypt.capi.CAPI.mongocrypt_t;
 import com.sun.jna.Pointer;
 import org.bson.BsonBinary;
 import org.bson.BsonDocument;
@@ -33,51 +34,51 @@ import java.security.SecureRandom;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Supplier;
 
-import static com.mongodb.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_ERROR;
-import static com.mongodb.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_FATAL;
-import static com.mongodb.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_INFO;
-import static com.mongodb.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_TRACE;
-import static com.mongodb.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_WARNING;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_crypt_shared_lib_version_string;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_datakey_init;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_decrypt_init;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_encrypt_init;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_explicit_decrypt_init;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_explicit_encrypt_expression_init;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_explicit_encrypt_init;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_new;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_rewrap_many_datakey_init;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_setopt_algorithm;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_setopt_algorithm_range;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_setopt_contention_factor;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_setopt_key_alt_name;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_setopt_key_encryption_key;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_setopt_key_id;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_setopt_key_material;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_ctx_setopt_query_type;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_destroy;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_init;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_is_crypto_available;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_new;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_aes_256_ctr;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_append_crypt_shared_lib_search_path;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_bypass_query_analysis;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_crypto_hook_sign_rsaes_pkcs1_v1_5;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_crypto_hooks;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_encrypted_field_config_map;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_kms_provider_aws;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_kms_provider_local;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_kms_providers;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_log_handler;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_schema_map;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_set_crypt_shared_lib_path_override;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_setopt_use_need_kms_credentials_state;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_code;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_destroy;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_message;
-import static com.mongodb.crypt.capi.CAPI.mongocrypt_status_new;
-import static com.mongodb.crypt.capi.CAPIHelper.toBinary;
+import static com.mongodb.internal.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_ERROR;
+import static com.mongodb.internal.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_FATAL;
+import static com.mongodb.internal.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_INFO;
+import static com.mongodb.internal.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_TRACE;
+import static com.mongodb.internal.crypt.capi.CAPI.MONGOCRYPT_LOG_LEVEL_WARNING;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_crypt_shared_lib_version_string;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_datakey_init;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_decrypt_init;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_encrypt_init;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_explicit_decrypt_init;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_explicit_encrypt_expression_init;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_explicit_encrypt_init;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_new;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_rewrap_many_datakey_init;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_setopt_algorithm;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_setopt_algorithm_range;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_setopt_contention_factor;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_setopt_key_alt_name;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_setopt_key_encryption_key;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_setopt_key_id;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_setopt_key_material;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_ctx_setopt_query_type;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_destroy;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_init;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_is_crypto_available;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_new;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_aes_256_ctr;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_append_crypt_shared_lib_search_path;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_bypass_query_analysis;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_crypto_hook_sign_rsaes_pkcs1_v1_5;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_crypto_hooks;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_encrypted_field_config_map;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_kms_provider_aws;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_kms_provider_local;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_kms_providers;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_log_handler;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_schema_map;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_set_crypt_shared_lib_path_override;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_setopt_use_need_kms_credentials_state;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_code;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_destroy;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_message;
+import static com.mongodb.internal.crypt.capi.CAPI.mongocrypt_status_new;
+import static com.mongodb.internal.crypt.capi.CAPIHelper.toBinary;
 import static org.bson.assertions.Assertions.isTrue;
 import static org.bson.assertions.Assertions.notNull;
