+++ mongodb-crypt/src/main/com/mongodb/internal/crypt/capi/MongoCryptOptions.java
@@ -15,9 +15,10 @@
  *
  */
 
-package com.mongodb.crypt.capi;
+package com.mongodb.internal.crypt.capi;
 
 import org.bson.BsonDocument;
+
 import java.util.List;
 import java.util.Map;
