+++ mongodb-crypt/src/main/com/mongodb/internal/crypt/capi/SLF4JLogger.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.mongodb.crypt.capi;
+package com.mongodb.internal.crypt.capi;
 
 import org.slf4j.LoggerFactory;
